added new files

+++ backend/routes/posts.js
@@ -1,41 +1,106 @@
 const express = require('express');
+const multer = require('multer');
 
 const Post = require('../models/post');
+const checkAuth = require('../middleware/check-auth');
 const router = express.Router()
 
-router.post("",(req,res,next)=>{
+const MIME_TYPE_MAP = {
+  'image/png': 'png',
+  'image/jpeg': 'jpg',
+  'image/jpg': 'jpg'
+}
+
+const storage = multer.diskStorage({
+  destination: (req,file,cb)=>{
+    const isValid = MIME_TYPE_MAP[file.mimetype];
+    let error = new Error("Invalid mime type");
+    if(isValid){
+      error=null;
+    }
+    cb(error, "backend/images"); //relative to server.js file
+  },
+  filename: (req,file,cb) => {
+    const name = file.originalname.toLowerCase().split(' ').join('-');
+    const extension = MIME_TYPE_MAP[file.mimetype];
+    cb(null,name+'-'+Date.now()+'.'+extension);
+  }
+});
+
+router.post("",
+  checkAuth,
+  multer({storage:storage}).single("image"),
+  (req,res,next)=>{
+  const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
-    content: req.body.content
+    content: req.body.content,
+    imagePath: url+"/images/"+req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost=>{
     res.status(201).json({
       message: "Post Added",
-      postId: createdPost._id
+      //Instead of PostID return a Post
+      //postId: createdPost._id
+      post: {
+        // id: createdPost._id,
+        // title: createdPost.title,
+        // content: createdPost.content,
+        //or
+        ...createdPost,
+        imagePath: createdPost.imagePath
+      }
     });
   });
 
 });
 
-router.put("/:id",(req,res,next)=>{
+router.put("/:id",checkAuth,multer({storage:storage}).single("image"),(req,res,next)=>{
+  let imagePath = req.body.imagePath
+  if(req.file){
+    const url = req.protocol + '://' + req.get("host");
+    imagePath = url+"/images/"+req.file.filename
+  }
   const post = new Post({
     _id:req.body.id,
     title:req.body.title,
-    content:req.body.content
-  })
-  Post.updateOne({_id:req.params.id},post).then(result=>{
-    console.log(result);
-    res.status(200).json({message:"Update Successful!"});
+    content:req.body.content,
+    imagePath: imagePath,
+    creator: req.userData.userId
+  });
+
+  Post.updateOne({_id:req.params.id, creator: req.userData.userId},post).then(result=>{
+    if(result.nModified>0){
+      res.status(200).json({message:"Update Successful!"});
+    } else{
+      res.status(401).json({message:"Unauthorized Access"});
+    }
+
   })
 })
 
 
 router.get("", (req, res, next) => {
-  Post.find()
+  const pageSize = +req.query.pagesize;//+ converts string to integer
+  const currentPage = +req.query.page;
+  const postQuery = Post.find();
+  let fetchedPosts;
+  if(pageSize && currentPage){
+    postQuery
+      .skip(pageSize*(currentPage - 1))
+      .limit(pageSize);
+  }
+  postQuery
   .then((documents)=>{
-    res.status(200).json({
-      message: 'Posts fetched unsuccesfully!',
-      posts: documents
+    fetchedPosts = documents;
+    return Post.count();
+    })
+    .then(count=>{
+      res.status(200).json({
+        message: 'Posts fetched succesfully!',
+        posts: fetchedPosts,
+        maxPosts: count
     });
   });
 
@@ -51,15 +116,15 @@ router.get("/:id", (req,res,next)=>{
   })
 })
 
-router.delete("/:id",(req,res,next)=>{
-  Post.deleteOne({_id:req.params.id})
+router.delete("/:id",checkAuth,(req,res,next)=>{
+  Post.deleteOne({_id:req.params.id, creator: req.userData.userId})
   .then(result=>{
-    console.log("Post deleted");
-    res.status(200).json({
-      message: "Post deleted"
-    });
+    if(result.n>0){
+      res.status(200).json({message:"Deletion Successful!"});
+    } else{
+      res.status(401).json({message:"Unauthorized Access"});
+    }
   });
-
 });
 
 module.exports = router;
